@@ -14,21 +14,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.9",
-    "@angular/cdk": "17.3.9",
-    "@angular/common": "17.3.9",
-    "@angular/compiler": "17.3.9",
-    "@angular/core": "17.3.9",
-    "@angular/forms": "17.3.9",
-    "@angular/material": "17.3.9",
-    "@angular/platform-browser": "17.3.9",
-    "@angular/platform-browser-dynamic": "17.3.9",
-    "@angular/router": "17.3.9",
+    "@angular/animations": "18.0.4",
+    "@angular/cdk": "18.0.4",
+    "@angular/common": "18.0.4",
+    "@angular/compiler": "18.0.4",
+    "@angular/core": "18.0.4",
+    "@angular/forms": "18.0.4",
+    "@angular/material": "18.0.4",
+    "@angular/platform-browser": "18.0.4",
+    "@angular/platform-browser-dynamic": "18.0.4",
+    "@angular/router": "18.0.4",
     "@aws-amplify/ui-angular": "^5.0.8",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "8.0.0",
     "@ngxs/store": "3.8.2",
-    "@nx/angular": "18.3.5",
+    "@nx/angular": "19.3.1",
     "@rx-angular/cdk": "17.0.0",
     "@rx-angular/template": "17.0.0",
     "ably": "^1.2.48",
@@ -42,24 +42,24 @@
     "zone.js": "0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.8",
-    "@angular-devkit/core": "17.3.8",
-    "@angular-devkit/schematics": "17.3.8",
-    "@angular-eslint/eslint-plugin": "17.3.0",
-    "@angular-eslint/eslint-plugin-template": "17.3.0",
-    "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "~17.3.0",
-    "@angular/compiler-cli": "17.3.9",
-    "@angular/language-service": "17.3.9",
-    "@nx/cypress": "18.3.5",
-    "@nx/eslint": "18.3.5",
-    "@nx/eslint-plugin": "18.3.5",
-    "@nx/jest": "18.3.5",
-    "@nx/js": "18.3.5",
-    "@nx/storybook": "18.3.5",
-    "@nx/web": "18.3.5",
-    "@nx/workspace": "18.3.5",
-    "@schematics/angular": "17.3.8",
+    "@angular-devkit/build-angular": "18.0.5",
+    "@angular-devkit/core": "18.0.5",
+    "@angular-devkit/schematics": "18.0.5",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
+    "@angular-eslint/template-parser": "18.0.1",
+    "@angular/cli": "~18.0.0",
+    "@angular/compiler-cli": "18.0.4",
+    "@angular/language-service": "18.0.4",
+    "@nx/cypress": "19.3.1",
+    "@nx/eslint": "19.3.1",
+    "@nx/eslint-plugin": "19.3.1",
+    "@nx/jest": "19.3.1",
+    "@nx/js": "19.3.1",
+    "@nx/storybook": "19.3.1",
+    "@nx/web": "19.3.1",
+    "@nx/workspace": "19.3.1",
+    "@schematics/angular": "18.0.5",
     "@storybook/addon-essentials": "7.5.1",
     "@storybook/angular": "7.6.0",
     "@storybook/core-server": "7.6.0",
@@ -67,18 +67,19 @@
     "@types/node": "18.19.10",
     "@typescript-eslint/eslint-plugin": "7.9.0",
     "@typescript-eslint/parser": "7.9.0",
+    "@typescript-eslint/utils": "^8.0.0-alpha.28",
     "autoprefixer": "^10.4.17",
-    "cypress": "^13.6.6",
+    "cypress": "13.9.0",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-cypress": "^2.15.1",
     "eslint-plugin-prefer-arrow-functions": "^3.2.4",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "jest-preset-angular": "14.0.4",
+    "jest-preset-angular": "14.1.1",
     "jsonc-eslint-parser": "^2.4.0",
-    "ng-packagr": "17.3.0",
-    "nx": "18.3.5",
+    "ng-packagr": "18.0.0",
+    "nx": "19.3.1",
     "postcss": "^8.4.33",
     "postcss-import": "~14.1.0",
     "postcss-preset-env": "~7.5.0",
